+++ src/Areas/Workbooks/Options/BaseWorkbooksOptions.cs
@@ -1,0 +1,9 @@
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT License.
+
+using System.Text.Json.Serialization;
+using AzureMcp.Options;
+
+namespace AzureMcp.Areas.Workbooks.Options;
+
+public class BaseWorkbooksOptions : GlobalOptions;